Trying to fix a bug(s?) opening a ruby file.
When you open a file with shebang, encoding, or frozen string comment it
sometimes gets hung up in font lock. The error you get is related to
backing up past the start of the file. This should address that.

Of note, I found out that the error message in question was not hitting
when debug-on-error is set not because the error was being swallowed
here, but because debug-ignored-errors was swallowing it. Good to know.