Make erm more resilient towards errors.
Still not perfect. Need to figure out odd encoding errors that seem to
want me to be in binary mode. Need to make it better about not bugging
me if I've opened and killed a bunch of buffers quickly. The errors
from killing of buffers in particular are annoying. Makes using macros
tedious.