Fixed bug in enh-ruby-up-sexp-test that required a sleep.
The gist was that enh-ruby-previous-indent-change was returning nil in
some cases. It never should and I'm defaulting to point-min for now.

This is really a problem with TESTING TEXT PROPERTIES IN ORDER TO
AVOID PARSING the file directly... but the text properties are coming
in asynchronously and might not be done yet. erm-wait-for-parse is a
bit of a lie in that it is dependent on erm-parsing-p and doesn't
actually wait until done (I think?).