Added string-should-indent-like-ruby and many tests.
This compares enh-ruby-mode against ruby-mode and expects them to be
the same. There are 2 new failed tests that I think are incorrect on
the ruby-mode side. I'll need to explore them further.

I'd also like to add tests for the various deep-indent options but
that can wait for now (pure test PRs welcome!).