Mostly cleanup and TODOs
I want to change the way deep-indent works for arrays / word arrays.
Namely, the current indentation is a bit off:

words = %w[ a
          b
          c
        ]
words = [ 'a',
          'b',
          'c'
        ]

when it should be more like:

words = %w[ a
            b
            c
          ]
words = [ 'a'
          'b',
          'c'
        ]

There should be at least as much indentation added as non-deep
indentation, which defaults to 2, not 1. The close brace should match
the opening brace, not the '%' part.

I can be talked out of the indent of 2 in favor of 1, but the
alignment should definitely be respected.