Don't depend on ruby-mode for show-paren support
We were delegating to ruby-mode's smie to get the show-paren-mode
behavior for "if"/"end" pairs, but that was causing smie to take over
indent-line-function as well.

Now erm implements its own advice around show-paren-data-function so
show-paren-mode works for "if"/"end" pairs.

Note that it is also possible to "fix" the ruby-mode dependency by
moving the smie loading code to the top of enh-ruby-mode setup so so
that erm indent-line-function overwrites the smie one. However, not
depending on ruby-mode is obviously better.

Using erm appears be much faster than smie as well (smie had noticable
lag for humongous "if"/"end" blocks).

Known differences/issues:
 - You only get show-paren highlighting when point is after "end" or
   on the first character of the opener (e.g. the "i" in "if"). I did
   this intentionally because I thought ruby-mode's behavior
   highlights things too often, and this behavior is more consistent
   with normal show-paren-mode.
 - There are some cases of invalid pairs that still highlight as valid
   due to enh-ruby-{backward,forward}-sexp bouncing to the "wrong"
   thing. I didn't attempt to fix this.