Temporary fix for forward-sexp-function being so heinous.
Dunno why, but when I'm working on ruby_parser in particular (very
large test files and the like), this change seems to seriously *bog*
everything down. Until I can get proper profiling up and running
against fixed data, I need to get this out of the way.

This does break the ability to use things like `sort-sexp` and the
like. It'll have to wait.