Fixed enh-ruby--changed-vars-with and generalized both all-* and changed-*.
Cribbing off of customize-rogue, I found out that you *need* to eval
the car of standard-value in order to compare it properly. This
normalizes all sorts of wackiness between '(...) (list ...) etc. It
was the only way to make enh-ruby-encoding-map not show up as changed
every time even tho I don't actually modify it.

Still... scary that I have to eval it. O_o

Also made it so they both use string-match-p so I can use regexps to
get subsets out.