Fix enh-ruby-toggle-block edge cases / fix tests
With a hash inside a brace-delimited block, `enh-ruby-toggle-block`
would mistake the hash for the block and give incorrect results. This
PR makes `enh-ruby-toggle-block` leverage text properties to determine
where the block, and tries to improve code clarity.

It also fixes skipped tests for `enh-ruby-toggle-block` by waiting for
erm to parse the buffer and for fontification to happen before running
expectations.

Closes #132