Remove faulty "error handling" from erm-ruby-get-process
if erm-parsing-p is t and erm-response is "" it would call error w/
an empty message. I also saw cases where erm-response was '"', which
just isn't information I can work with anyway... so throwing in all
cases is a better solution because it'll clean up the old process and
get a new one started smoothly.